+++ applications/example/init/clientServerEvents.js
@@ -1,0 +1,11 @@
+if (application.isWorker) {
+
+  application.frontend.on('test', function(data) {
+    console.dir(data);
+  });
+
+  setInterval(function() {
+    application.frontend.send('test', { field: 'value' });
+  }, 10000);
+
+}