+++ applications/example/server/examples/events/connect.ws/get.js
@@ -1,5 +1,4 @@
 module.exports = function(client, callback) {
-
   var connection = client.websocket.accept();
   if (connection) {
     connection.send('Hello world');
@@ -11,5 +10,4 @@ module.exports = function(client, callback) {
     //});
   }
   callback();
-
 };