+++ applications/example/server/examples/impress.rpc/get.js
@@ -1,8 +1,4 @@
 module.exports = function(client, callback) {
-  if (client.rpc) client.rpc.accept({
-    simple: '/examples/simple',
-    tools: '/examples/tools',
-    memory: '/examples/memory'
-  });
+  if (client.rpc) client.rpc.accept();
   callback();
 };