+++ deploy/centos7x64.sh
@@ -5,15 +5,15 @@ yum -y install wget
 yum -y install psmisc
 yum -y groupinstall "Development Tools"
 cd /usr/src
-wget http://nodejs.org/dist/v4.1.0/node-v4.1.0.tar.gz
-tar zxf node-v4.1.0.tar.gz
-rm -f ./node-v4.1.0.tar.gz
-cd node-v4.1.0
+wget http://nodejs.org/dist/v4.1.1/node-v4.1.1.tar.gz
+tar zxf node-v4.1.1.tar.gz
+rm -f ./node-v4.1.1.tar.gz
+cd node-v4.1.1
 ./configure
 make
 make install
 cd ~
-rm -rf /usr/src/node-v4.1.0
+rm -rf /usr/src/node-v4.1.1
 ln -s /usr/local/bin/node /bin
 ln -s /usr/local/bin/npm /bin
 cat >/etc/yum.repos.d/mongodb.repo <<EOL